pyproject.toml
[project]
name = "jobrag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jobrag"]

[project.scripts]
jobrag = "jobrag.cli:main"